+++ frontend/src/app/onboarding/page.tsx
@@ -2,10 +2,8 @@
 
 import OnboardingForm from '@/components/form/OnboardingForm';
 import { redirect } from 'next/navigation';
-import { Skeleton, VStack } from '@chakra-ui/react';
+import SkeletonArray from '@/components/skeleton/SkeletonArray';
 import { useUserData } from '../../hooks/useUserData';
-
-const skeletonArray = new Array(10).fill(0);
 
 export default function Account() {
   const { data, isLoading } = useUserData();
@@ -15,13 +13,7 @@ export default function Account() {
   }
 
   if (isLoading) {
-    return (
-      <VStack spacing={6} align="stretch">
-        {skeletonArray.map(() => (
-          <Skeleton h={10} style={{ borderRadius: '0.375rem' }} />
-        ))}
-      </VStack>
-    );
+    return <SkeletonArray />;
   }
 
   return <OnboardingForm />;
